ISPN-12067 HTTP/2 framing error for invalid requests
